fix(api): report a programming related illegal arg error when...
...resourcePath is null or empty

This usually indicates that a validation rule is missing for the path
input value in the interface or somewhere in the call stack.
We could add it here, but given that this is a generic place we can only
add a generic error that might not help identify which property is
missing or erroneous.